Work on job refactoring
